Move TableSchema code into its own module
This is part of the query model, but it isn't a Node subclass and the
more we can split this module up the better.
