+++ tests/unit/query_model/test_nodes.py
@@ -33,19 +33,6 @@ EVENTS_SCHEMA = TableSchema(
 )
 
 
-# TEST TABLESCHEMA
-#
-
-
-def test_table_schema_equality():
-    t1 = TableSchema(i=Column(int))
-    t2 = TableSchema(i=Column(int))
-    t3 = TableSchema(j=Column(int))
-    assert t1 == t2
-    assert t1 != t3
-    assert t1 != "a fish"
-
-
 # TEST BASIC QUERY MODEL PROPERTIES
 #
